requirements.txt
asgiref==3.3.4
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==3.2
django-cors-headers==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.8
flake8==5.0.4
flake8-docstrings==1.7.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pep8-naming==0.13.3
pillow==11.0.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.9.0
pytest==7.1.3
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.5.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.4.1
tomli==2.0.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
django-filter==23.1
django-storages==1.8